Basic Character Movement + Began Terrain Coloring based on height/slope/etc
Created first person camera to walk around with high speeds for now (reduce later?). Currently can walk while in the air and can't jump which would be good future things to fix.

Terrain can be colored with code based on specific factors, but this can't be accessed through a UI yet.